% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_abcd2vis}
\alias{gv_abcd2vis}
\title{Compute complex visibility based on ABCD principle}
\usage{
gv_abcd2vis(ii, kx = as.numeric(c(4)), sx = as.numeric(c(4)),
  fx = as.numeric(c(1)))
}
\arguments{
\item{ii}{ABCD}

\item{kx}{P2VM. By default, c(0,2,2,0)}

\item{sx}{P2VM. By default, c(4,4,4,4)}

\item{fx}{Total flux in ABCD}
}
\description{
Compute complex visibility based on ABCD principle
}

