% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_analytic}
\alias{gv_analytic}
\title{Compute the analytic signal}
\usage{
gv_analytic(sig, fc = NULL, debug = F)
}
\arguments{
\item{sig}{The signal}

\item{fc}{Optional. Cut-off frequency of a low-pass filter.}
}
\description{
Compute the analytic signal
}

