% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_blorder}
\alias{gv_blorder}
\title{Remap an arbitrary baseline arrangement to the default}
\usage{
gv_blorder(M)
}
\arguments{
\item{M}{telescope matrix of size 6x4}
}
\description{
A default baseline arrangment in GRAVITY is given as:
T1-T2, T1-T3, T1-T4, T2-T3, T2-T4, T3-T4
}

