% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_calib_intopd}
\alias{gv_calib_intopd}
\title{Compute the close loop OPD between the FT, SC and MET fringes}
\usage{
gv_calib_intopd(fits, v2pms, drk, cal, pol = 1, debug = F, high.res = F,
  ka = c(0, 2, 2, 0), hdu_sc = "imaging_data_sc",
  hdu_ft = "imaging_data_ft", hdu_met = "metrology", hdu_ddl = "fddl")
}
\arguments{
\item{fits}{Filename of the GRAVITY merged FITS}

\item{v2pms}{A list of V2PM coefficients to reduce SC fringes}

\item{drk}{Dark pixels and read noise from gv_sci_Nfits2pix().}

\item{cal}{A list of calibration coefficients}

\item{p}{(optional) Polarization states to process.}

\item{lam}{(optional) A vector of mean wavelengths for the FT, SC and MET fringes in microns.}
}
\description{
Compute the close loop OPD between the FT, SC and MET fringes
}

