% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_ft_pix2vis}
\alias{gv_ft_pix2vis}
\title{Computes the complex visibilty of fringes}
\usage{
gv_ft_pix2vis(flux, p2vm, blseq = NULL)
}
\arguments{
\item{flux}{Vector of pixel values}

\item{p2vm}{List of P2VM matrices}
}
\description{
Computes the complex visibilty of the FT fringes in a pixel array
}

