% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_ft_readmfits}
\alias{gv_ft_readmfits}
\title{Read GD/PD of FT fringes for a given FITS file}
\usage{
gv_ft_readmfits(fits, hdu = "imaging_data_ft")
}
\arguments{
\item{fits}{A FITS file to be read}

\item{hdu}{Optional. Name of HDU in the FITS file to be read}
}
\description{
Read GD/PD of FT fringes for a given FITS file
}

