% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_met_compdcv}
\alias{gv_met_compdcv}
\title{Compute the differential OPD of the metrology path}
\usage{
gv_met_compdcv(cv)
}
\arguments{
\item{cv}{A list of OPD per telescope}
}
\description{
Compute the differential OPD of the metrology path
}

