% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_met_procmfits}
\alias{gv_met_procmfits}
\title{Compute PD of MET fringes}
\usage{
gv_met_procmfits(fits, sel = 1:4, Nps = 4, Ncpu = 1, ka = c(0, 2, 2, 0),
  hdu = "metrology")
}
\arguments{
\item{fits}{FITS file to process}

\item{sel}{Optional. Columns within a HDU/volt table/column to process}

\item{Nps}{Optional. Number of shifts per phase estimates (must be multiple of 4)}

\item{Ncpu}{Optional. Use multiple processors}

\item{ka}{Optional. Phase shift calibration}

\item{hdu}{Optional. The name of the HDU to process}
}
\description{
Compute PD of MET fringes
}

