% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_met_volt2vis}
\alias{gv_met_volt2vis}
\title{Compute complex visibility for a time series of voltages}
\usage{
gv_met_volt2vis(volt, N = 4, abcd = 1:4, ka = c(0, 2, 2, 0))
}
\arguments{
\item{volt}{Vector of diode voltages}

\item{N}{Number of shifts per phase estimates (must be multiple of 4)}

\item{abcd}{Initial phase shift}

\item{ka}{Phase shift calibration}
}
\description{
Compute complex visibility for a time series of voltages
}

