% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_plot_cp}
\alias{gv_plot_cp}
\title{Plot closure phase}
\usage{
gv_plot_cp(cp, x = NULL, d = 1, e_cp = NULL, epsfile = NULL, ...)
}
\arguments{
\item{cp}{Matrix of closure phases. One triangle per row.}
}
\description{
Plot closure phase
}

