% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_plot_errorbars}
\alias{gv_plot_errorbars}
\title{Plot errorbars}
\usage{
gv_plot_errorbars(x, y, ex = NULL, ey = NULL, w = 0.05, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{ex}{Error bar for x}

\item{ey}{Error bar for y}
}
\description{
Plot errorbars
}

