% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_plot_opd}
\alias{gv_plot_opd}
\title{Plot GD or GD+PD}
\usage{
gv_plot_opd(opd, x = NULL, d = 1, offset = 0, e_opd = NULL,
  epsfile = NULL, noleg = F, ...)
}
\arguments{
\item{opd}{Data to be plot. One set per row.}
}
\description{
Plot GD or GD+PD
}

