% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_readcsv}
\alias{gv_readcsv}
\title{Read and write a custom formatted CSV file}
\usage{
gv_readcsv(filein)
}
\arguments{
\item{filein}{Path and name of the csv file to be read}

\item{fileout}{Path and name of the csv file to be written}

\item{arr}{An array to be written to fileout}

\item{append}{Should an existing file be overwritten? By default, no.}
}
\value{
A named list of arrays
}
\description{
Such a file stores multiple 1D/2D arrays in a csv format. Each array is
preceeded with a header describing its name, row and colum. This format is
currently used for storing gvspc calibration data.
}

