% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_readfits}
\alias{gv_readfits}
\title{FITS file reader}
\usage{
gv_readfits(fits_name, hdu_name = "", col_name = "")
}
\arguments{
\item{fits_name}{FITS file. If a vector is given, only the first is read.}

\item{hdu_name}{Name of the HDU to be read. Case-insensitive. If not provided, the Primary HDU is read.}

\item{col_name}{Name of the column of an extended HDU to be read. This is ignored if the HDU is an image.}
}
\value{
A numeric array.
}
\description{
Reads a specific image HDU or extended binary table of a FITS files.
This reader function is tailor made for GRAVITY FITS format but can be used
for generic FITS files. However, not all data types are supported.
}
\details{
Extended ASCII table and complex data types are currently not supported.
}

