% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_checkcal}
\alias{gv_sci_checkcal}
\title{Check SCI detector calibration text file}
\usage{
gv_sci_checkcal(file_cal, file_fits, file_dark = NULL, file_eps = NULL,
  rev = 3, hdu = "imaging_data_spe")
}
\arguments{
\item{file_cal}{The calibration file}

\item{file_fits}{A FITS file containing images from the SCI detector}

\item{file_eps}{Optional. An eps file to export plot. By default, no.}

\item{rev}{Optional. For development purpose.}

\item{hdu}{Optional. The name of the HDU containing the images. For primary HDU, set it to an empty string.}
}
\description{
Cross-check the indices in the calibration file, which are shared between
R and C/C++ functions, with an actual FITS file.
}
\details{
2014-05-29, Yitping - Initial rev.
2014-06-03, Yitping - To support the GRAVITY merged FITS format
}

