% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_fits2vis}
\alias{gv_sci_fits2vis}
\title{Computes the complex visibility of the fringes}
\usage{
gv_sci_fits2vis(fits, v2pms, dark, cal, p = 1, hdu = "imaging_data_spe",
  blseq = NULL)
}
\arguments{
\item{fits}{A GRAVITY FITS file}

\item{cal}{A list of calibration coefficients}

\item{v2pms}{A list of V2PM coefficients}

\item{dark}{Dark pixels and read noise from gv_sci_Nfits2pix().}

\item{p}{Optional. Polarization states to process.}

\item{hdu}{Optional. The hdu name that contains the SCI inteferograms.}

\item{blseq}{Optional. Baseline order and OPD polarity. See gv_blorder().}
}
\value{
A 3D cube of complex visibility vs. spectral channel vs. baseline vs. number of image frames.
}
\description{
Computes the complex visibilty of the SCI fringes in a given FITS file.
}

