% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_Nfits2pix}
\alias{gv_sci_Nfits2pix}
\title{Read multiple fits}
\usage{
gv_sci_Nfits2pix(fits, cal, hdu = "")
}
\arguments{
\item{fits}{FITS files}

\item{cal}{List of calibration matrices}

\item{hdu}{Name of the HDU table to be read}
}
\description{
Read multiple fits
}

