% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_procmfits}
\alias{gv_sci_procmfits}
\title{Compute GD/PD of SC fringes for a given FITS file}
\usage{
gv_sci_procmfits(fits, v2pms, drk, cal, pol = 1, hdu = "imaging_data_sc")
}
\arguments{
\item{fits}{FITS file to be processed}

\item{v2pms}{List of v2pm matrices}

\item{drk}{Dark pixels and read noise from gv_sci_Nfits2pix().}

\item{cal}{A list of calibration coefficients}

\item{p}{Optional. Polarization states to process.}

\item{hdu}{Optional. The hdu name that contains the SCI inteferograms.}
}
\description{
Compute GD/PD of SC fringes for a given FITS file
}

