% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_readcal}
\alias{gv_sci_readcal}
\title{Parse SCI detector calibration text file}
\usage{
gv_sci_readcal(file_cal, n_io = gv_const()$sci_io_out)
}
\arguments{
\item{file_cal}{The calibration file}
}
\description{
Read the SCI detector calibration file and format the content properly for
usage with other GRAVITY R functions, e.g. to convert the SCI raw images to
pixel arrays, etc.
}

