% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_readp2vm}
\alias{gv_sci_readp2vm}
\title{Parse SCI P2VM text file}
\usage{
gv_sci_readp2vm(file_p2vm, n_io = gv_const()$sci_io_out)
}
\arguments{
\item{file_p2vm}{The P2VM file}
}
\description{
Read the SCI detector P2VM file and format the content properly for usage
with other GRAVITY R functions, e.g. to convert the SCI pixel arrays to OPDs,
etc.
}

