% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_sci_readv2pm}
\alias{gv_sci_readv2pm}
\title{Parse SCI V2PM text file}
\usage{
gv_sci_readv2pm(file)
}
\arguments{
\item{file}{P2VM file}
}
\description{
Read the SCI detector V2PM file and format the content properly for usage
with other GRAVITY R functions, e.g. to convert the SCI pixel arrays to OPDs,
etc.
}

