% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_telmat}
\alias{gv_telmat}
\title{Telescope matrix}
\usage{
gv_telmat(map = gv_const()$sci_telarr, flip = T)
}
\arguments{
\item{map}{Telescope vector}
}
\description{
Maps a vector of telescopes to a vector of baselines.
By default, the order of telescope pair follows the arrangement as seen on
the SCI detector.
}

