% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_vis2cp}
\alias{gv_vis2cp}
\title{Compute closure phases or OPDs}
\usage{
gv_vis2cp(vis, cpi = rbind(c(1, 4, 2), c(1, 5, 3), c(4, 6, 5)))
}
\arguments{
\item{vis}{Complex visiblity of a vector of baselines.}

\item{opd}{OPD of a vector of baselines.}
}
\value{
A vector of closure phases or OPDs
}
\description{
By default the vector baseline whose closure phases or OPDs are to be computed
are arranged in the default GRAVITY sequence, i.e. T1-T2, T1-T3, and so on.
}

