% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_vis2opd}
\alias{gv_vis2opd}
\title{Estimate OPD from complex visibility.}
\usage{
gv_vis2opd(cv, R = 1, lam = 1, type = "pd", debug = F)
}
\arguments{
\item{cv}{Complex visibility vector}

\item{R}{Optional. Spectral resolution of the spectrometer.}

\item{lam}{Optional. Wavelength in the middle of the spectrum.}

\item{type}{Optional. Selects how OPD is computed.}
}
\value{
Estimated OPD.
}
\description{
Esimate OPD from complex visibility of spectrally dispersed fringes.
}
\details{
Selects between the following:
\itemize{
  \item{pd}{Phase delay (default)}
  \item{gd}{Group delay}
  \item{op}{OPD (experimental)}
}
}

