% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_weighted.mean}
\alias{gv_weighted.mean}
\title{Compute weighted mean and variance of column vectors/matrices}
\usage{
gv_weighted.mean(x, v)
}
\arguments{
\item{x}{Values}

\item{v}{Variance of each value}
}
\description{
Compute weighted mean and variance of column vectors/matrices
}

