% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_writefits_img}
\alias{gv_writefits_img}
\title{FITS image writer}
\usage{
gv_writefits_img(img, fits_name, hdu_name = "")
}
\description{
Writes a vector, matrix or 3D array to a FITS file as an image.
The data is written to the primary HDU.
}

