% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_writefits_tbl}
\alias{gv_writefits_tbl}
\title{FITS table writer}
\usage{
gv_writefits_tbl(dat, fits_name, hdu_name = "gRavity", units = "")
}
\description{
Writes a named list to a FITS file as a binary table.
The data is written to an extended HDU. Each member of the list is written
as a separate column in the table.
}
\details{
Supports only numeric and integer data type for now.
}

