% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_ft_fits2vis}
\alias{gv_ft_fits2vis}
\title{Computes the complex visibility of the fringes}
\usage{
gv_ft_fits2vis(fits, p2vm, dark = 0, p = 1, hdu = "imaging_data_ft",
  blseq = NULL, ma = 5)
}
\description{
Computes the complex visibilty of the FT fringes in a given FITS file
}

