% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gv_solvels}
\alias{gv_solvels}
\title{Solve for x in an Ax=b linear system}
\usage{
gv_solvels(rA, rb, rw = numeric())
}
\description{
The solution for x is inv(A)%*%b where inv(A) is computed by means of the
singular value decomposition method.
}

